+++ buildSrc/src/main/kotlin/Versions.kt
@@ -1,6 +1,6 @@
 object Versions {
     const val KOTLIN = "1.8.10"
-    const val ATOMIC_FU = "0.18.5"
+    const val ATOMIC_FU = "0.20.0"
     const val ANDROID_GRADLE_PLUGIN = "7.3.1"
     const val JETPACK_COMPOSE_COMPILER = "1.4.3"
     const val JETPACK_COMPOSE_RUNTIME = "1.3.3"
@@ -10,8 +10,8 @@ object Versions {
     const val JETPACK_COMPOSE_MATERIAL_3 = "1.1.0-alpha07"
     const val COIL = "2.0.0-rc03"
     const val KTLINT = "11.2.0"
-    const val KOTLINX_SERIALIZATION = "1.4.1"
-    const val KOTLINX_COROUTINES = "1.6.3"
+    const val KOTLINX_SERIALIZATION = "1.5.0"
+    const val KOTLINX_COROUTINES = "1.6.4"
     const val KTOR = "2.0.3"
     const val KOTLIN_WRAPPERS_EXTENSIONS = "1.0.1-pre.459"
     const val ANDROIDX_LIFECYCLE = "2.6.0-rc01"